EID-602 Move tests to package with conforming name.
