--- verify-matching-service-test-tool/src/main/resources/verify/special-name-case-with-non-latin-name.json
@@ -1,44 +1,0 @@
-  {
-    "hashedPid": "8fbfdd035a92daf890673d9ff8f50def3aad66ea42026c581bd73735250ba706",
-    "levelOfAssurance": "LEVEL_1",
-    "matchId": "default-match-id",
-    "matchingDataset": {
-      "addresses": [
-        {
-          "internationalPostCode": "GB1 2PF",
-          "lines": [
-            "10 George Street"
-          ],
-          "postCode": "GB1 2PF",
-          "uprn": "833F1187-9F33-A7E27B3F211E",
-          "verified": true
-        }
-      ],
-      "dateOfBirth": {
-        "value": "1980-05-24",
-        "verified": true
-      },
-      "firstName": {
-        "value": "Georgios",
-        "nonLatinScriptValue": "Γεώργιος",
-        "verified": true
-      },
-      "middleName": {
-        "value": "Xanthi",
-        "nonLatinScriptValue": "Ξάνθη",
-        "verified": true
-      },
-      "gender": {
-        "value": "MALE",
-        "verified": true
-      },
-
-      "surnames": [
-        {
-          "value": "Panathinaikos",
-          "nonLatinScriptValue": "Παναθηναϊκός",
-          "verified": true
-        }
-      ]
-    }
-  }
