--- verify-matching-service-test-tool/src/main/resources/verify/special-name-case.json
@@ -1,43 +1,0 @@
-  {
-    "hashedPid": "8fbfdd035a92daf890673d9ff8f50def3aad66ea42026c581bd73735250ba706",
-    "levelOfAssurance": "LEVEL_1",
-    "matchId": "default-match-id",
-    "matchingDataset": {
-      "addresses": [
-        {
-          "internationalPostCode": "GB1 2PF",
-          "lines": [
-            "10 George Street"
-          ],
-          "postCode": "GB1 2PF",
-          "uprn": "833F1187-9F33-A7E27B3F211E",
-          "verified": true
-        }
-      ],
-      "dateOfBirth": {
-        "value": "1980-05-24",
-        "verified": true
-      },
-      "firstName": {
-        "value": "Šarlota",
-        "nonLatinScriptValue": "Šarlota",
-        "verified": true
-      },
-      "gender": {
-        "value": "FEMALE",
-        "verified": true
-      },
-      "middleNames": {
-        "value": "Snježana",
-        "nonLatinScriptValue": "Snježana",
-        "verified": true
-      },
-      "surnames": [
-        {
-          "value": "Jõey",
-          "nonLatinScriptValue": "Jõey",
-          "verified": true
-        }
-      ]
-    }
-  }
